refactor: move venue module in a folder so that we can add its components in the next commit
