+++ src/features/home/components/index.ts
@@ -2,4 +2,4 @@ export { HomeHeader } from './headers/HomeHeader'
 export { BusinessModule } from './modules/BusinessModule'
 export { ExclusivityModule } from './modules/exclusivity/ExclusivityModule'
 export { OffersModule } from './modules/OffersModule'
-export { VenuesModule } from './modules/VenuesModule'
+export { VenuesModule } from './modules/venues/VenuesModule'
